setup/pyproject.toml
[project]
name = "setup"
version = "0.1.0"
description = "Code for setting up Prefect Server and its workers"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "prefect[docker]>=3.4.20",
    "python-dotenv>=1.1.1",
]
